requirements.txt
Flask
flask-cors
opencv-python-headless
numpy
tensorflow
Pillow
gunicorn
mlflow
